🚚 Move poll core UI into features/poll (RAL-1291)
- Move components/poll/** (desktop-poll, mobile-poll, manage-poll,
  voting-form, poll-context-provider, etc.) to features/poll/components/
- Move poll-context.tsx (+test), vote-icon/, event-card, event-meta and
  layouts/poll-layout.tsx to features/poll/components/
- Extract language-selector.tsx to components/ (domain-agnostic, only
  used by settings preferences)
- Remove 17 entries from the biome.json migration allowlist; keep
  manage-poll.tsx exempted for its pre-existing @/app duplicate-dialog
  import that the old components entry was masking
- Repoint use-csv-exporter dayjs/datetime exemptions (biome.json and
  no-raw-datetime-formatting.grit) to the new path
