+++ apps/web/biome.json
@@ -121,29 +121,13 @@
     {
       "includes": [
         "src/components/add-to-calendar-button.tsx",
-        "src/components/event-card.tsx",
-        "src/components/layouts/poll-layout.tsx",
         "src/components/logo.tsx",
-        "src/components/poll-context.tsx",
-        "src/components/poll/desktop-poll.tsx",
-        "src/components/poll/desktop-poll/participant-row-form.tsx",
-        "src/components/poll/desktop-poll/participant-row.tsx",
-        "src/components/poll/manage-poll.tsx",
-        "src/components/poll/manage-poll/schedule-poll-dialog.tsx",
-        "src/components/poll/mobile-poll.tsx",
-        "src/components/poll/mobile-poll/poll-option.tsx",
-        "src/components/poll/mobile-poll/poll-options.tsx",
-        "src/components/poll/notification-toggle.tsx",
-        "src/components/poll/poll-branding.tsx",
-        "src/components/poll/poll-context-provider.tsx",
-        "src/components/poll/poll-footer.tsx",
-        "src/components/poll/responsive-results.tsx",
-        "src/components/poll/voting-form.tsx",
         "src/components/pro-badge.tsx",
         "src/components/upgrade-button.tsx",
         "src/components/user-provider.tsx",
         "src/components/visibility.tsx",
         "src/features/billing/components/pay-wall-dialog.tsx",
+        "src/features/poll/components/manage-poll.tsx",
         "src/lib/auth.ts",
         "src/lib/feature-flags/config.ts",
         "src/lib/safe-action/server.ts"
@@ -175,7 +159,7 @@
         "src/trpc/routers/index.ts",
         "src/app/api/private/utils/time-slots.ts",
         "src/lib/utils/date-time-utils.ts",
-        "src/components/poll/manage-poll/use-csv-exporter.ts",
+        "src/features/poll/components/manage-poll/use-csv-exporter.ts",
         "src/test/setup.ts"
       ],
       "linter": {
