+++ apps/web/src/app/[locale]/(space)/settings/preferences/components/language-preference.tsx
@@ -14,7 +14,7 @@ import { ArrowUpRight } from "lucide-react";
 import Link from "next/link";
 import { useForm } from "react-hook-form";
 import * as z from "zod";
-import { LanguageSelect } from "@/components/poll/language-selector";
+import { LanguageSelect } from "@/components/language-selector";
 import { Trans } from "@/i18n/client";
 import { useLocale } from "@/lib/locale/client";
 import { trpc } from "@/trpc/client";
