+++ apps/web/src/app/[locale]/invite/[urlId]/invite-page-loader.tsx
@@ -1,6 +1,6 @@
 "use client";
 import dynamic from "next/dynamic";
-import { PollBrandingFromContext } from "@/components/poll/poll-branding";
+import { PollBrandingFromContext } from "@/features/poll/components/poll-branding";
 
 const InvitePage = dynamic(
   () => import("./invite-page").then((mod) => mod.InvitePage),
