+++ apps/web/src/app/[locale]/invite/[urlId]/invite-page.tsx
@@ -2,12 +2,12 @@
 import { Alert, AlertDescription } from "@rallly/ui/alert";
 import { ArrowUpRightIcon, CrownIcon } from "lucide-react";
 import Link from "next/link";
-import { EventCard } from "@/components/event-card";
-import { PollFooter } from "@/components/poll/poll-footer";
-import { ResponsiveResults } from "@/components/poll/responsive-results";
-import { VotingForm } from "@/components/poll/voting-form";
 import { usePoll } from "@/features/poll/client";
 import Discussion from "@/features/poll/components/discussion";
+import { EventCard } from "@/features/poll/components/event-card";
+import { PollFooter } from "@/features/poll/components/poll-footer";
+import { ResponsiveResults } from "@/features/poll/components/responsive-results";
+import { VotingForm } from "@/features/poll/components/voting-form";
 import { Trans } from "@/i18n/client";
 import { trpc } from "@/trpc/client";
