+++ apps/web/src/features/poll/components/manage-poll/delete-poll-dialog.tsx
