+++ apps/web/src/features/poll/components/mobile-poll/grouped-options.tsx
