+++ apps/web/src/features/poll/components/mobile-poll/poll-option.tsx
