+++ apps/web/src/features/poll/components/poll-branding.tsx
