+++ apps/web/src/features/poll/components/poll-footer.tsx
