+++ apps/web/src/features/poll/components/score-summary.tsx
@@ -4,9 +4,8 @@ import { AnimatePresence } from "motion/react";
 import * as m from "motion/react-m";
 import * as React from "react";
 import { usePrevious } from "react-use";
-
-import { usePoll } from "@/components/poll-context";
 import { IfScoresVisible } from "@/components/visibility";
+import { usePoll } from "@/features/poll/components/poll-context";
 
 export interface PopularityScoreProps {
   yesScore: number;
