+++ apps/web/src/features/poll/components/types.ts
