+++ apps/web/src/features/poll/components/vote-icon.tsx
@@ -2,10 +2,10 @@ import type { VoteType } from "@rallly/database";
 import { cn } from "@rallly/ui";
 import type { VariantProps } from "class-variance-authority";
 import { cva } from "class-variance-authority";
-import { IfNeedBeIcon } from "@/components/vote-icon/if-need-be-icon";
-import { NoIcon } from "@/components/vote-icon/no-icon";
-import { PendingIcon } from "@/components/vote-icon/pending-icon";
-import { YesIcon } from "@/components/vote-icon/yes-icon";
+import { IfNeedBeIcon } from "@/features/poll/components/vote-icon/if-need-be-icon";
+import { NoIcon } from "@/features/poll/components/vote-icon/no-icon";
+import { PendingIcon } from "@/features/poll/components/vote-icon/pending-icon";
+import { YesIcon } from "@/features/poll/components/vote-icon/yes-icon";
 
 const iconVariants = cva("", {
   variants: {
