+++ apps/web/src/features/poll/components/vote-icon/if-need-be-icon.tsx
