+++ apps/web/src/features/poll/components/vote-icon/no-icon.tsx
