+++ apps/web/src/features/poll/components/vote-icon/pending-icon.tsx
