+++ apps/web/src/features/poll/components/vote-icon/yes-icon.tsx
