+++ apps/web/src/features/poll/components/vote-selector.tsx
