+++ apps/web/src/features/poll/components/you-avatar.tsx
